+++ tools/urt/tools/quake3/q3map2/writebsp.c
@@ -164,7 +164,7 @@ void EmitLeaf( node_t *node ){
 	for ( b = node->brushlist; b; b = b->next )
 	{
 		/* something is corrupting brushes */
-		if ( (int) b < 256 ) {
+		if ( (size_t) b < 256 ) {
 			Sys_FPrintf( SYS_WRN, "WARNING: Node brush list corrupted (0x%08X)\n", b );
 			break;
 		}
